backend/requirements.txt
fastapi
uvicorn[standard]
playwright
beautifulsoup4
google-genai
python-dotenv
httpx
loguru
orjson
tenacity
apscheduler
numpy<2
